(ml5717) Some code refactoring + added lineage reference metadata
